* Windows compatibility * Visual Studio 2005 solution and project files * Reformatted README * Static mesh loader removed (grossly incompatible with VS C compiler) * Have to restrain to OpenGL 1.1 compatibility * Glib C code is not compatible with VS C compiler, need to rewrite

+++ src/common/c_vectors.h
@@ -137,7 +137,7 @@ static inline float C_vec2_square_len(c_vec2_t p)
 
 static inline float C_vec2_len(c_vec2_t p)
 {
-    return sqrt(C_vec2_square_len(p));
+    return (float)sqrt(C_vec2_square_len(p));
 }
 
 static inline float C_vec3_square_len(c_vec3_t p)
@@ -147,7 +147,7 @@ static inline float C_vec3_square_len(c_vec3_t p)
 
 static inline float C_vec3_len(c_vec3_t p)
 {
-    return sqrt(C_vec3_square_len(p));
+    return (float)sqrt(C_vec3_square_len(p));
 }
 
 /******************************************************************************\
